requirements.txt
labgrid
requests
pytest
attrs
fastapi
uvicorn
